Add final test score check to TakeTest test

+++ database/migrations/2026_04_28_110844_create_test_attempts_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('test_id')->constrained()->cascadeOnDelete();
-            // $table->integer('score')->nullable();
-            // $table->integer('max_score')->nullable();
-            $table->float('score_percentage')
+
+            // 5, 2 means 5 total digits, 2 after decimal point
+            $table->decimal('score_percentage', 5, 2)
                 ->nullable()
                 ->comment('The percentage score of the test attempt, calculated as (score / max_score) * 100');
             $table->text('comment')->nullable();
